pyproject.toml
[project]
name = "streamlet"
version = "2.3.0"
description = "A data pipeline with simple to configure periodic tasks build on celery."
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.5.3,<6.0.0",
    "PyYAML>=6.0.2,<7.0.0",
    "simplejson>=3.20.0,<4.0.0",
    "voluptuous>=0.15.2,<1.0.0",
]

[dependency-groups]
modules = [
    "psycopg[pool,binary]>=3.2.10,<4.0.0",
    "requests>=2.32.5,<3.0.0",
    "opensearch-py>=3.0.0,<4.0.0",
    "opentelemetry-sdk>=1.37.0,<2.0.0",
    "opentelemetry-exporter-otlp>=1.37.0,<2.0.0",
    # "confluent-kafka>=2.9.0,<2.10.0",
]
dev = [
    "black>=25.9.0,<26.0.0",
    "celery-types>=0.23.0,<1.0.0",
    "isort>=6.0.1, <7.0.0",
    "pylint>=3.3.8,<4.0.0",
    "pytest>=8.4.2,<9.0.0",
    "pytest-cov>=7.0.0,<8.0.0",
    "pytest-httpserver>=1.1.3,<2.0.0",
    "pytest-xdist>=3.8.0,<4.0.0",
    "testcontainers[redis,postgres,opensearch]>=4.13.0,<5.0.0",
]

[tool.uv]
default-groups = ["modules"]

[tool.black]
line_length = 100

[tool.isort]
line_length = 100
